requirements.txt
celery==3.1.26.post2
Django==3.0.5
django-allauth==0.49.0
django-celery==3.3.1
django-crispy-forms==1.13.0
django-filter==21.1
django-import-export==2.7.1
django-localflavor==3.1
django-social-share==2.2.1
django-widget-tweaks==1.4.8
numpy==1.21.4
pandas==1.3.5
Pillow==8.4.0
requests==2.26.0
scikit-learn==1.0.2
scipy==1.8.0
sklearn==0.0
textblob==0.17.1
